pkg/internal/flags/flags.go: moving to internal package, combining into one struct
