+++ pkg/helm/flags/flag.go
@@ -15,21 +15,19 @@
 package flags
 
 import (
-	"github.com/operator-framework/operator-sdk/pkg/flags"
+	"github.com/operator-framework/operator-sdk/pkg/internal/flags"
 	"github.com/spf13/pflag"
 )
 
 // HelmOperatorFlags - Options to be used by a helm operator
 type HelmOperatorFlags struct {
-	flags.ReconcilePeriodFlag
-	flags.WatchesFileFlag
+	flags.WatchFlags
 }
 
 // AddTo - Add the helm operator flags to the the flagset
 // helpTextPrefix will allow you add a prefix to default help text. Joined by a space.
 func AddTo(flagSet *pflag.FlagSet, helpTextPrefix ...string) *HelmOperatorFlags {
 	hof := &HelmOperatorFlags{}
-	hof.ReconcilePeriodFlag.AddTo(flagSet, helpTextPrefix...)
-	hof.WatchesFileFlag.AddTo(flagSet, helpTextPrefix...)
+	hof.WatchFlags.AddTo(flagSet, helpTextPrefix...)
 	return hof
 }
